refactor: setError 함수 utils로 분리
