+++ src/apis/profile.ts
@@ -3,7 +3,8 @@ import withToken from '@/apis/withToken';
 import { BadgeKeyType } from '@/types/challengeTypes';
 import { ApiSuccessResponse } from '@/types';
 import { getCookie } from '@/utils';
-import { handleError, setError } from '@/utils/api';
+import { handleError } from '@/utils/api';
+import { setError } from '@/utils/error';
 
 const profileApiUrl = process.env.NEXT_PUBLIC_API + '/profile';
